cleaned up file structure for plugins
