+++ utils/rollup-babel-options.mjs
@@ -1,6 +1,32 @@
 /** @typedef {import('@rollup/plugin-babel').RollupBabelInputPluginOptions} RollupBabelInputPluginOptions */
 
 /**
+ * The ES6 options for babel(...) plugin.
+ *
+ * @param {string} buildType - Only 'debug' requires special handling so far.
+ * @returns {RollupBabelInputPluginOptions} The babel options.
+ */
+const moduleOptions = buildType => ({
+    babelHelpers: 'bundled',
+    babelrc: false,
+    comments: buildType === 'debug',
+    compact: false,
+    minified: false,
+    presets: [
+        [
+            '@babel/preset-env', {
+                bugfixes: true,
+                loose: true,
+                modules: false,
+                targets: {
+                    esmodules: true
+                }
+            }
+        ]
+    ]
+});
+
+/**
  * The ES5 options for babel(...) plugin.
  *
  * @param {string} buildType - Only 'debug' requires special handling so far.
@@ -29,4 +55,4 @@ const es5Options = buildType => ({
     ]
 });
 
-export { es5Options };
+export { moduleOptions, es5Options };
